requirements.txt
Flask
python-dotenv
pandas
openpyxl
pyyaml
numpy
gunicorn
werkzeug
requests
beautifulsoup4
lxml
